Polish codes
Change Log:
- Post filtering test codes are separated from Post CRUD test codes
- Refactoring
